requirements.txt
pandas==2.2.2
matplotlib==3.9.0
sqlalchemy==2.0.30
pymysql==1.1.1
python-dotenv==1.0.1
